revert relationship with datasetlock - just use id
I originally changed to have a relationship between the
pendingworkflowinovcation and a datasetlock, but this complicates
creating and removing the lock, which happens in different transactions
than the ones that create the pendingworkflowinovcation. Further, the
lock applies for the entire workflow, whereas the
pendingworkflowinvocation is related to one step.
Just having the pendingworkflowinvocation add/remove the dataset lock id
as a long avoids the issue. I assume a relation between the two objects
could be made to work, but I'm not sure it is worth the effort.